Bug-fix Matrix: Matrix now initialises all values to be zero in the constructor.
Documentation: Removed documentation in Matrix3D that was untrue, added
extra documentation. 

Clean-up: Removed the second checkInteractionWithBoundaries from
DPMBase::computeOneTimeStep, as this does not seem to do anything for any of
the current boundaries. Also removed one relevant todo for this.

Clean-up: Removed a todo in BaseBoundary that was a question, and put the
answer in the documentation. Also re-did formating in BaseBoundary.

Clean-up: In FluxBoundary, removed commented out fprintf in
checkBoundaryAfterParticleMoved.

Clean-up: In BaseHandler, removed a todo.


git-svn-id: https://svn.mercurydpm.org/SourceCode/Trunk@2761 b3179fc6-a625-46cb-afb8-a2a51a81216d