BrokenMaserMinimalExample - A *very* deconstructed version of my Maser drivers that doesn't work in parallel.
db-minimal.pars - parameters file

Example usage:
  mpirun -n 2 ./BrokenMaserMinimalExample db-minimal.pars

The simulation is initiated by placing particles in a region using a
CubeInsertionBoundary, which is removed immediately. After a number of
timesteps, a SubcriticalMaserBoundary is added and activated. Two timesteps
later, the simulation hangs, with the last messages from the
DPMBase::computeOneTimeStep() loop being

[Process: 0]: about to call checkAndDuplicatePeriodicParticles()
[Process: 1]: about to call performGhostParticleUpdate()



git-svn-id: https://svn.mercurydpm.org/SourceCode/Trunk@2804 b3179fc6-a625-46cb-afb8-a2a51a81216d