Hertzian2DSelfTest - A broken usage of the Hertzian contact model in 2D with fixed particles
A particle that collides with a fixed particle will then have a nan position, 
suggesting that the 'infinite mass' of fixed particles isn't being dealt with
properly. The problem isn't there for collisions between two movables. The
problem seems to occur for 3D as well, but I'm not sure if the elastic modulus
has an appropriate value.


git-svn-id: https://svn.mercurydpm.org/SourceCode/Trunk@2905 b3179fc6-a625-46cb-afb8-a2a51a81216d