Fix a couple of compiler compatibility issues
DPMBase.cc - fix an error introduced by commit 3075, which replaced
  *interactionHandler.getObject(i)

with 
  interactionHandler->getObject(i)

The original code actually does the operator precedence as
  *(interactionHandler.getObject(i))

so the star has lower precedence than the dot. For some reason this compiled
fine for me on GCC 7.2 and c++14, but didn't work on clang.

Particles/SuperQuadric.cc - Replace math.h with cmath

math.h seems to be broken on GCC 7.2 (at least on the installation at DAMTP)


git-svn-id: https://svn.mercurydpm.org/SourceCode/Trunk@3079 b3179fc6-a625-46cb-afb8-a2a51a81216d