bug-fix: parallel code
Changed the definition of communication complexity.
Previously:
1 = particle is located on one side of a domain
2 = particle is located on a rib of a domain
3 = particle is located on a corner of a domain

New definition just uses the local grid coordinate to identify with
which domains the particle is communicating with

complexity = i + 3*j + 9*k + 13 
Here 13 is the center (away from all boundaries)

This to avoid ambiguity when the particle moves from one side to another
which can happen in very rare cases. In those cases the old definition
of communication complexity did not change.



git-svn-id: https://svn.mercurydpm.org/SourceCode/Trunk@3215 b3179fc6-a625-46cb-afb8-a2a51a81216d