Merging back in ParallelMaserDebugging, 3/3: Domain.cc: if the other particle in the interaction is not found, the code now continues without copying the interaction. Also updated documentation.
ParticleHandler.cc: in addObject, the periodic complexity is set to a vector
of length 0. 

PeriodicBoundaryHandler.cc: updated comments.

BaseParticle.cc: in copy-constructor, the periodic complexity is not copied
anymore.

SubcriticalMaserBoundaryTEST: 
- extra feature: copy the flow-particles to the front a few times when
activating the maser.
- rewrote activateMaser so that there are no problems with periodic
complexity anymore: making particles maser-particles should be done after
copying the bottom and possibly flow-particles.

DeletionBoundary.cc: initialised variables to 0 in constructor, to prevent
segfault.

ParticleVtkWriter: added a 'particleMustBeWritten' method because the
if-statement in the SphericalParticleVtkWriter was becoming cumbersome, and
this is easier for testing out things. Also cleanup.

SphericalParticleVtkWriter: now using particleMustBeWritten, also cleanup.


git-svn-id: https://svn.mercurydpm.org/SourceCode/Trunk@3239 b3179fc6-a625-46cb-afb8-a2a51a81216d