This commit deals with groups of walls. We want contacts with a group of walls to persist when the contact point switches smoothly from contact to the other.
This is mostly necessary to improve teh behaviour of TriangleWalls read in from STL files: currently, particles feel a bump where two TriangleWall faces touch, and the history parameters get lost as well.

This needs a change in BaseWall::getInteractionWith: if a particle has multiple contacts with a group of walls, and if the contact areas of these contacts overlap, we keep only the biggest of the overlapping contacts.

git-svn-id: https://svn.mercurydpm.org/SourceCode/Trunk@4862 b3179fc6-a625-46cb-afb8-a2a51a81216d