ArcWall - improved documentation/comments and unit tests
The comments in ArcWall.cc and ArcWall.h have been improved, and the logic in
ArcWall::getDistanceAndNormal() has been clarified.

This change creates ArcWallUnitTest, which covers some cases in which the pipe
is aligned with the z axis (e.g. in a 2D simulation).

A new constant constants::degree has been defined as a shortcut for pi/180.

git-svn-id: https://svn.mercurydpm.org/SourceCode/Trunk@6422 b3179fc6-a625-46cb-afb8-a2a51a81216d