BUGFIX::MDPM-22 Implement the logger everywhere, i.e. all cout and cerr should go through the logger - Resolved. All cout() and cerr() functions replaced.
DPMBase.cc DPMBase.h MercuryBase.h MercuryBase.cc
added new private variable nToWrite_ to fix a bug caused by virtual functions and the DPMBase as well as the MercuryBase class. The DPMBase function was always called instead of the MercuryBase class. In this way HGrid was missing and restarttests failed

git-svn-id: https://svn.mercurydpm.org/SourceCode/Trunk@6594 b3179fc6-a625-46cb-afb8-a2a51a81216d