fix: More consistent neighbor interaction checks within the membrane class
Membrane: Made handling of neighboring interactions more consistent with respect to contact history; it is now checked after force calculation (which is the case for the contact type already). Previously it depended on the ID if a contact was detected. This may have an effect for particles who go from one triangle to another one.

Approved-by: Timo Plath
Approved-by: Thomas Weinhart