feat: Converted Loadstatistics.m into a python version Loadstatistics.py
Merged in feature/Python_LoadStatistics (pull request #51)

Loadstatistics was converted into a Python version to enable ease of MercuryCG use when working in Python (e.g. with GrainLearning)

Currently the python code works for .ene, .stat and .restart files. the .data files are currently not working and I will only implement them on request as the basic functionality is already guaranteed.

Loadstatistics.py naming is following the PEP8 python convention we agreed on

when all matrices are transposed beforehand they lead to the same plot if we transpose again. So transposing does not lead to different results.

Anyway, transposed all matrices in the python version of LoadStatistics to make it match the MATLAB version.

Added a new DotDict class to be able to call the Loadstatistics data dictionaries by dot.notation.

Approved-by: Anthony Thornton
Approved-by: Thomas Weinhart