Merged in mergeMercuryLab (pull request #76)
MergeMercuryLab (Bring in improves to liquid bridge model)

* MERGE :: ML change to mercurydpm version

* Remove the ML drivers dir

* bugfix: fixed several compiler warnings and added override statements to several functions

Merged in bugfix/fix_compiler_warnings (pull request #75)

Approved-by: Timo Plath
Approved-by: Anthony Thornton
* Merged in cleanup/update-licenses (pull request #78)

updated licenses

* refactor: updated licenses

* Merged master into cleanup/update-licenses

Approved-by: Anthony Thornton
Approved-by: Thomas Weinhart
* Merged in feature/create-gitattributes (pull request #79)

Added a .gitattributes file which forces LF line endings

* feat: added a .git attributes file which forces LF line endings exceptions that have to be explicitly defined.

In this way no one has to do git config --global core.autocrlf=true.

* Merged master into feature/create-gitattributes

Approved-by: Anthony Thornton
Approved-by: Mitchel Post
Approved-by: Thomas Weinhart
* refactor: renamed Tutorials in documentation

Merged in cleanup/rename-tutorials-in-doc (pull request #71)

Approved-by: Thomas Weinhart
Approved-by: Thomas Weinhart
* refactor: Add .vscode to .gitignore.

This patch adds the directory of local settings of Visual Studio Code
to the .gitignore file.

Approved-by: Thomas Weinhart
Approved-by: Timo Plath
* feat: Add getMaximumNumberOfOMPThreads().

Gets the maximum number of OpenMP threads if MERCURYDPM_USE_OMP is enabled.

Add an example for helpers::getMaximumNumberOfOMPThreads() to the
OpenMP tutorial.

Merged in get-max-omp-threads (pull request #89)

Approved-by: Timo Plath
Approved-by: Thomas Weinhart
* build: Add `MercuryDPM_KERNEL_ONLY` as advanced CMake option.

Disables drivers, tools, xballs and tests (~75% of the build).
It can speed up development/deployment.

Merged in cmake-kernel-only (pull request #88)

Approved-by: Timo Plath
Approved-by: Thomas Weinhart
* feat: Add {set,get}WritePythonFileForVTKVisualisation().

This feature allows the user to enable/disable the generation of the .py file.
It is enabled by default to keep the original behaviour.

Rename `shouldWritePythonFileForVTKVisualisation_` to
`forceWritePythonFileForVTKVisualisation_` to be consistent with
`forceWriteOutputFiles`.

Merged in disable-vtk-py (pull request #86)

Approved-by: Timo Plath
Approved-by: Thomas Weinhart

Approved-by: Timo Plath
Approved-by: Thomas Weinhart