build: fixed nearly all of the compiler warnings when building MercuryDPM.
Removed unused variables in FMM

Fixed a few warnings, mostly unused variables

changed error warning in MixedSpecies.h and Species.h in the copyInto function.
gcc created a lot of error messages showing that the pointer to s is nullptr. Which makes sense since we first check for it being nullptr and then use it. So I removed the usage of s->getName().

fixed an MPI related bug.
getSmallestParticle() should not be used with MPI, therefore I changed it to getSmallestInteractionRadius(). same for largest particle.

Merged in cleanup/fixCompilerWarnings (pull request #93)

Approved-by: Timo Plath
Approved-by: Igor Ostanin