chore: added Mercury random number generator to the PSD Class
introduced new RNG object into the PSD class to draw random numbers for the drawSample() function used in particle insertion. Added a new function as well which lets you set a fixed seed for the RNG object to make selfTests reproducible.

Added PSDSelfTest.cpp again, as it seems to be deleted by a former commit.

All Selftests including a PSD are updated to have a fixed seed of zero.

Replaced the PSD random number generator by the internal mercury RNG class

Created a new function in the PSD class PSD::setFixedSeed(int seed) to give reproducible results for SelfTests.

Replaced all SelfTestData for the tests using the PSD class

The new default seed for the RNG in the PSD class is zero.

Converted the PSD in DistributionSelfTest.cpp to a more sensible one where no zero size particles can be created; These led to delay in the execution of the selftest.

Fixed a bug in MasterMake.cmake introduced in commit 2329d55. The error lead to SelfTests not being run, because the SelfTest gold data searched for included the path of the file.
Added a second layer of check in fpdiff.py such that SelfTests will fail if a SelfTest gold data file is not found.
also added an error message to the fpdiff.py details_stream which catches the error when an empty input file is used for gold data (i.e. it cannot find the gold data, or it does not exist)

fixes Issue MDPM-433

Merged in feature/MDPM-433-implement-the-mercury-rng-to-th (pull request #59)

Approved-by: Anthony Thornton
Approved-by: Thomas Weinhart