bugfix: fixed the version dropdown menu to correctly display all versions in the online documentation.
fixed wrong naming of the javascript file.

fixed a bug where a regular expression did delete the last letter of a version while searching for the trailing slash. There was a dot which was not escaped, so it took any character.

Merged in cleanup/DocumentationVersionSelectorFix (pull request #116)

Approved-by: Thomas Weinhart
merged with 1 approve since it was urgent to close all PR's due to a large file in git history