bugfix: fixed temporarily failing selfTests due to undefined behaviour for adhesive speices. Hopefully this resolve the sometimes failing StressStrainControlBoundrySelfTest.
fixed an invalid read that would go out of scope and was called by reference. The fix links it to the object itself and will make sure it does not go out of scope.

Fixed a condition that depended on an uninitialised variable. maxInteractionDistance was never initialised and produced random values. This led to problems when the .restart file was read into StressStrainControlBoundarySelfTest. When adding the species to the speciesHandler it tried to setMaxInteractionDistance() based on an undefined maxInteractionDistance_

Merged in cleanup/fix-selfTests-in-Release (pull request #109)

Approved-by: Thomas Weinhart
We (Timo and Thomas) discussed these changes together.