feat: added Eigen as a submodule to MercuryDPM
MercuryDPM_EIGEN is by default off and can be turned on via CMake.

added EigenUnitTest.cpp

moved EigenUnitTest.cpp to its own directory and disabled it if MercuryDPM_EIGEN is not enabled. This prevents the test from failing if we do not include Eigen

Discussed with Thomas to set Eigen default to OFF

included Helpers to EigenUnitTest.cpp and now check for the last timestep gridpoint values.

added success message for Eigen cloning and initialisation

Merged in feature/MDPM-472-add-eigen (pull request #85)

Approved-by: Timo Plath
Approved-by: Thomas Weinhart