build: changed Python3 to be REQUIRED and Git to be required for oomph-lib
Changed CMake behaviour to avoid in the future the problem we had with the pipelines:

Python3 is now a required package
Git is now a required package if oomph-lib is being installed via CMake

updated dockerfile and image to include python3-dev. This will include PythonLibs

removed PythonLibs 3 do not need to be required in CMake. These are only header files to compile python libraries and are not a requirement for fpdiff.py