feat: PrintWallTimeMixin class to estimate wall time of a simulation
* Alternative estimate based on last time

* [DPMBase] normalize whitespace

* Slightly simplify ifdefs by using a guard clause

* More on PrintWallTimeMixin

* Dead-reckoned estimate for when the simulation will complete (based on
  total time elapsed so far, so will be inaccurate if the number of
  particles varies).

* Advanced formatting to support displaying durations that are hours
  long.

Test plan:

    ./PrintWallTimeMixinDemo 1  # test milliseconds get formatted properly
    ./PrintWallTimeMixinDemo 10 -tmax 100  # minutes
    ./PrintWallTimeMixinDemo 10 -tmax 1000  # hours

* PrintWallTimeMixin

A mixin is a small class that contributes an amount of functionality or
behaviour, often small and optional. In this case, PrintWallTimeMixin
overrides the printTime method to print not just t and tmax, but also a
timestamp of the real time when that line was logged. This helps track
the progress of a simulation

* deleted the PrintWallTimeMixin object in PrintWallTimeMixinDemo.cpp.
* Documented PrintWallTimeMixin.cc and .h

Merged in feature/printWallTimeMixin (pull request #147)